+++ tests/testthat/test-obs_opts.R
@@ -7,7 +7,6 @@ test_that("obs_opts returns expected default values", {
   expect_true(result$week_effect)
   expect_equal(result$week_length, 7L)
   expect_equal(result$scale, Normal(mean = 1, sd = 0))
-  expect_equal(result$accumulate, 0)
   expect_true(result$likelihood)
   expect_false(result$return_likelihood)
 })