+++ tests/testthat/test-rt_opts.R
@@ -31,10 +31,10 @@ test_that("rt_opts handles custom inputs correctly", {
   expect_equal(result$gp_on, "R0")
 })
 
-test_that("rt_opts warns when pop is passed as numeric", {
-  lifecycle::expect_deprecated(
+test_that("rt_opts errors when pop is passed as numeric", {
+  expect_error(
     rt_opts(pop = 1000),
-    "The `pop` argument of `rt_opts\\(\\)` must be a `<dist_spec>` as of EpiNow2 1.7.0."
+    "must be a `<dist_spec>`"
   )
 })